requirements.txt
fastapi
uvicorn
pydantic
httpx
bs4
supabase
dotenv-python